requirements.txt
streamlit
pandas==1.5.3
sounddevice
scipy>=1.10,<1.11.4
python-dotenv
pycaret
openai
langfuse
streamlit-webrtc
av